requirements.txt
asn1crypto==0.24.0
awscli==1.16.19
boto3==1.9.9
botocore==1.12.9
certifi==2018.8.24
chardet==3.0.4
colorama==0.3.9
docutils==0.14
gitdb2==2.0.5
GitPython==2.1.11
idna==2.7
jmespath==0.9.3
lxml==4.2.5
mypy==0.630
mypy-extensions==0.4.1
nose==1.3.7
pyasn1==0.4.4
python-dateutil==2.7.3
python-lambda-local==0.1.6
pytz==2018.5
PyYAML==3.13
requests==2.20.0
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
smmap2==2.0.5
typed-ast==1.1.0
urllib3==1.23
wincertstore==0.2
